+++ src/tasks/get.py
@@ -9,12 +9,11 @@ import httpx
 import ijson
 import orjson
 import polars as pl
-from httpx import Client, HTTPStatusError, TimeoutException, get
+from httpx import Client, get
 from lxml import etree, html
 from prefect.transactions import transaction
 from tenacity import (
     retry,
-    retry_if_exception_type,
     stop_after_attempt,
     wait_exponential,
 )
@@ -25,6 +24,7 @@ from src.config import (
     DECP_USE_CACHE,
     HTTP_CLIENT,
     HTTP_HEADERS,
+    LOG_LEVEL,
     RESOURCE_CACHE_DIR,
     SIRENE_UNITES_LEGALES_URL,
     DecpFormat,
@@ -40,24 +40,23 @@ from src.tasks.transform import prepare_unites_legales
 from src.tasks.utils import (
     full_resource_name,
     gen_artifact_row,
+    get_logger,
     stream_replace_bytestring,
 )
 
 
-@retry(
-    stop=stop_after_attempt(3),
-    wait=wait_exponential(multiplier=1, min=1, max=10),
-    retry=retry_if_exception_type(httpx.HTTPError),  # On ne retry que sur erreur http
-)
+@retry(stop=stop_after_attempt(3), wait=wait_exponential(multiplier=1, min=1, max=20))
 def stream_get(url: str, chunk_size=1024**2):  # chunk_size en octets (1 Mo par défaut)
+    logger = get_logger(level=LOG_LEVEL)
+
     if url.startswith("http"):
         try:
             with HTTP_CLIENT.stream(
-                "GET", url, headers=HTTP_HEADERS, follow_redirects=True
+                "GET", url, headers=HTTP_HEADERS, follow_redirects=True, timeout=20
             ) as response:
                 yield from response.iter_bytes(chunk_size)
         except httpx.TooManyRedirects:
-            print(f"⛔️ Erreur 429 Too Many Requests pour {url}")
+            logger.error(f"⛔️ Erreur 429 Too Many Requests pour {url}")
             return
 
     else:
@@ -70,8 +69,10 @@ def stream_get(url: str, chunk_size=1024**2):  # chunk_size en octets (1 Mo par 
 def get_resource(
     r: dict, resources_artifact: list[dict] | list
 ) -> tuple[pl.LazyFrame | None, DecpFormat | None]:
+    logger = get_logger(level=LOG_LEVEL)
+
     if DECP_PROCESSING_PUBLISH is False:
-        print(f"➡️  {full_resource_name(r)}")
+        logger.info(f"➡️  {full_resource_name(r)}")
 
     output_path = DATA_DIR / "get" / r["filename"]
     output_path.parent.mkdir(exist_ok=True, parents=True)
@@ -88,9 +89,9 @@ def get_resource(
             fields, decp_format = xml_stream_to_parquet(
                 url, output_path, fix_chars=True
             )
-            print(f"♻️  {full_resource_name(r)} nettoyé et traité")
+            logger.info(f"♻️  {full_resource_name(r)} nettoyé et traité")
     else:
-        print(f"▶️  Format de fichier non supporté : {full_resource_name(r)}")
+        logger.warning(f"▶️  Format de fichier non supporté : {full_resource_name(r)}")
         return None, None
 
     if decp_format is None:
@@ -124,13 +125,15 @@ def get_resource(
 
 
 def find_json_decp_format(chunk, decp_formats, resource: dict):
+    logger = get_logger(level=LOG_LEVEL)
+
     for decp_format in decp_formats:
         decp_format.coroutine_ijson.send(chunk)
         if len(decp_format.liste_marches_ijson) > 0:
             # Le parser a trouvé au moins un marché correspondant à ce format, donc on a
             # trouvé le bon format.
             return decp_format
-    print(
+    logger.warning(
         f"⚠️  Pas de match trouvé parmis les schémas passés : {full_resource_name(resource)}"
     )
     return None
@@ -139,6 +142,8 @@ def find_json_decp_format(chunk, decp_formats, resource: dict):
 def json_stream_to_parquet(
     url: str, output_path: Path, resource: dict
 ) -> tuple[set, DecpFormat or None]:
+    logger = get_logger(level=LOG_LEVEL)
+
     decp_format_2019 = DecpFormat("DECP 2019", SCHEMA_MARCHE_2019, "marches")
     decp_format_2022 = DecpFormat("DECP 2022", SCHEMA_MARCHE_2022, "marches.marche")
     decp_formats = [decp_format_2019, decp_format_2022]
@@ -162,7 +167,6 @@ def json_stream_to_parquet(
 
     # Le dataset AWS scraping a pas mal de bugs de backslash
     if "/68caf6b135f19236a4f37a32/" in url or "/aws/" in url:
-        print("Remplacements spécifiques pour AWS...")
         stream_replace_iter = stream_replace_bytestring(
             stream_replace_bytestring(
                 stream_replace_bytestring(stream_replace_iter, rb"(\\\\\\)", rb"\\"),
@@ -177,7 +181,7 @@ def json_stream_to_parquet(
     try:
         chunk = next(stream_replace_iter)
     except StopIteration:
-        print(f"⚠️  Flux vide pour {url}")
+        logger.error(f"⚠️  Flux vide pour {url}")
         return set(), None
 
     decp_format = find_json_decp_format(chunk, decp_formats, resource)
@@ -370,6 +374,8 @@ def norm_titulaire(titulaire: dict):
 
 
 def get_etablissements() -> pl.LazyFrame:
+    logger = get_logger(level=LOG_LEVEL)
+
     schema = {
         "siret": pl.String,
         "codeCommuneEtablissement": pl.String,
@@ -382,7 +388,6 @@ def get_etablissements() -> pl.LazyFrame:
     }
 
     columns = list(schema.keys())
-    print(columns)
 
     base_url = "https://files.data.gouv.fr/geo-sirene/last/dep/"
     htmlpage: str = get(base_url).text
@@ -397,22 +402,19 @@ def get_etablissements() -> pl.LazyFrame:
             hrefs.append(base_url + href)
 
     # Fonction de traitement pour un fichier
+    @retry(
+        stop=stop_after_attempt(4), wait=wait_exponential(multiplier=1, min=1, max=20)
+    )
     def get_process_file(_href: str):
-        print(_href.split("/")[-1])
-        try:
-            response = http_client.get(
-                _href, headers=HTTP_HEADERS, timeout=20
-            ).raise_for_status()
-        except (HTTPStatusError, TimeoutException) as err:
-            print(err)
-            print("Nouvel essai...")
-            response = http_client.get(
-                _href, headers=HTTP_HEADERS, timeout=20
-            ).raise_for_status()
+        response = http_client.get(
+            _href, headers=HTTP_HEADERS, timeout=30
+        ).raise_for_status()
 
         content = response.content
         lff = pl.scan_csv(content, schema_overrides=schema)
         lff = lff.select(columns)
+        logger.info(_href.split("/")[-1], "OK")
+
         return lff
 
     # Traitement en parrallèle avec 8 threads
@@ -424,18 +426,20 @@ def get_etablissements() -> pl.LazyFrame:
                 lf = future.result()
                 lfs.append(lf)
             except Exception as e:
-                print(f"Error processing file: {e}")
+                logger.info(f"Error processing file: {e}")
 
-    print("Concaténation...")
+    logger.info("Concaténation...")
     lf_etablissements: pl.LazyFrame = pl.concat(lfs)
     return lf_etablissements
 
 
 def get_insee_cog_data(url, schema_overrides, columns) -> pl.DataFrame:
+    logger = get_logger(level=LOG_LEVEL)
+
     try:
         df_insee = pl.read_csv(url, schema_overrides=schema_overrides, columns=columns)
     except ConnectionResetError:
-        print("Connection error, retrying in 2 seconds...")
+        logger.error("Connection error, retrying in 2 seconds...")
         sleep(2)
         df_insee = get_insee_cog_data(
             url, schema_overrides=schema_overrides, columns=columns
@@ -446,6 +450,8 @@ def get_insee_cog_data(url, schema_overrides, columns) -> pl.DataFrame:
 def get_clean(
     resource, resources_artifact: list, available_parquet_files: set
 ) -> pl.DataFrame or None:
+    logger = get_logger(level=LOG_LEVEL)
+
     with transaction():
         checksum = resource["checksum"]
         parquet_path = RESOURCE_CACHE_DIR / f"{checksum}"
@@ -470,12 +476,14 @@ def get_clean(
                 return None
         else:
             # Le fichier parquet est déjà disponible pour ce checksum
-            print(f"👍 Ressource déjà en cache : {resource['dataset_code']}")
+            logger.debug(f"👍 Ressource déjà en cache : {resource['dataset_code']}")
             return parquet_path.with_suffix(".parquet")
 
 
 def get_unite_legales(processed_parquet_path):
-    print("Téléchargement des données unité légales et sélection des colonnes...")
+    logger = get_logger(level=LOG_LEVEL)
+
+    logger.info("Téléchargement des données unité légales et sélection des colonnes...")
     (
         pl.scan_parquet(SIRENE_UNITES_LEGALES_URL)
         .pipe(prepare_unites_legales)