+++ tests/test_clean.py
@@ -104,7 +104,7 @@ def test_clean_titulaires():
         ),
     }
     lf_2019 = pl.LazyFrame(data_2019, schema=schema_2019)
-    result_2019 = clean_titulaires(lf_2019, decp_format_2019).collect()
+    result_2019 = clean_titulaires(lf_2019, decp_format_2019, "titulaires").collect()
 
     titulaires_0 = result_2019["titulaires"][0]
     assert len(titulaires_0) == 2
@@ -141,7 +141,7 @@ def test_clean_titulaires():
         ),
     }
     lf_2022 = pl.LazyFrame(data_2022, schema=schema_2022)
-    result_2022 = clean_titulaires(lf_2022, decp_format_2022).collect()
+    result_2022 = clean_titulaires(lf_2022, decp_format_2022, "titulaires").collect()
 
     titulaires_0 = result_2022["titulaires"][0]
     assert len(titulaires_0) == 1
@@ -256,7 +256,7 @@ def test_clean_decp():
 
     # Test with DECP 2019
     decp_format_2019 = DecpFormat("DECP 2019", SCHEMA_MARCHE_2019, "marches")
-    df_result: pl.DataFrame = clean_decp(lf, decp_format_2019).collect()
+    df_result: pl.DataFrame = clean_decp(lf, decp_format_2019).sort("id").collect()
 
     # Check id cleaning
     assert df_result.filter(pl.col("id") == "id_1").height > 0